scripts/analyze_sentiment/requirements.txt
# Function dependencies, for example:
# package>=version
praw
beautifulsoup4
google-cloud-storage
google-cloud-language